Allow users to lock their files, require rollups's files to be locked
This commit adds "locked" Boolean property to files. When a file is
locked, it cannot be overwritten even by its owner until it expires.
The main purpose of this is to ensure that block data stored by rollup
inside Sharded Storage is stored for a certain period, and even rollup
itself cannot delete it later. I.e. we require the

As a side-effect, we make this feature avaliable to other users as well.
Any user can lock a file it creates, and prevent it from being modified
even by the user itself until the file expires. This is optional, and
regular users are not required to lock their files.